+++ tests/Services/ModelDriverTest.php
@@ -10,6 +10,7 @@
  */
 use Infuse\Application;
 use Pulsar\Driver\DatabaseDriver;
+use Pulsar\ErrorStack;
 use Pulsar\Model;
 use Pulsar\Services\ModelDriver;
 
@@ -23,10 +24,17 @@ class ModelDriverTest extends PHPUnit_Framework_TestCase
             ],
         ];
         $app = new Application($config);
+        $errorStack = new ErrorStack($app);
+        $app['errors'] = function () use ($errorStack) {
+            return $errorStack;
+        };
         $service = new ModelDriver($app);
         $this->assertInstanceOf(DatabaseDriver::class, Model::getDriver());
 
         $driver = $service($app);
         $this->assertInstanceOf(DatabaseDriver::class, $driver);
+
+        $model = new TestModel();
+        $this->assertEquals($errorStack, $model->getErrors());
     }
 }